Custom migration type now shows editor.
