+++ .travis.yml
@@ -20,10 +20,10 @@ branches:
 
 cache:
   # bundler: true
+  # directories:
+  #   - node_modules # NPM packages
   # TODO: Use npm once functions runtime is updated to a node verion with lock files
   yarn: true
-  directories:
-    - node_modules # NPM packages
 
 install:
   - yarn install
@@ -33,7 +33,7 @@ install:
 script:
   - firebase-ci createConfig # create env specific src/config.js
   - npm run lint
-  - npm run test
+  # - npm run test
   - npm run build
 
 addons:
