+++ src/routes/MigrationTemplate/components/MigrationTemplateAction/MigrationTemplateAction.enhancer.js
@@ -1,0 +1,5 @@
+import { compose } from 'redux'
+import { withHandlers } from 'recompose'
+import { withNotifications } from 'modules/notification'
+
+export default compose(withNotifications, withHandlers({}))
