+++ src/routes/MigrationTemplate/components/MigrationTemplateAction/index.js
@@ -1,0 +1,3 @@
+import MigrationTemplateAction from './MigrationTemplateAction'
+
+export default MigrationTemplateAction
